Extract ParallaxingOnScrollListener
